+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Fixed
+
+- Handle syntax errors from `ruff>=0.8.0`.
+
 ## 1.5.0 (2024-08-27)
 
 ### Added